fixed eslint config

+++ eslint.config.js
@@ -3,7 +3,7 @@ export default [
     files: ["**/*.js"],
     rules: {
       semi: "error",
-      quotes: ["error", "double"],
+      quotes: ["error", "single"],
     },
   },
 ];
